Backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vitória Emanuely"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.8"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.10.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-django = "^4.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
